switch to Script component on the facebook pixel example
